Fix signed-unsigned comparison
Bug: 35901901
Test: compilation yields no warning

+++ libril/ril_service.cpp
@@ -2998,7 +2998,7 @@ void fillCellIdentityFromVoiceRegStateResponseString(CellIdentity &cellIdentity,
     RIL_CellIdentity_v16 rilCellIdentity;
     int32_t *tmp = (int32_t*)&rilCellIdentity;
 
-    for (int i = 0; i < sizeof(RIL_CellIdentity_v16)/sizeof(int32_t); i++) {
+    for (size_t i = 0; i < sizeof(RIL_CellIdentity_v16)/sizeof(int32_t); i++) {
         tmp[i] = -1;
     }
 
@@ -3059,7 +3059,7 @@ void fillCellIdentityFromDataRegStateResponseString(CellIdentity &cellIdentity,
     RIL_CellIdentity_v16 rilCellIdentity;
     int32_t *tmp = (int32_t*)&rilCellIdentity;
 
-    for (int i = 0; i < sizeof(RIL_CellIdentity_v16)/sizeof(int32_t); i++) {
+    for (size_t i = 0; i < sizeof(RIL_CellIdentity_v16)/sizeof(int32_t); i++) {
         tmp[i] = -1;
     }
